refactor: 피드백 반영
- Results 생성자 변경
- Bridge test 추가
- 패키지 분리
