+++ src/main/java/domain/ladder/Bridge.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.Arrays;
