+++ src/main/java/domain/user/UserName.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.user;
 
 import java.util.regex.Pattern;
