+++ src/test/java/domain/game/ResultsTest.java
@@ -1,12 +1,12 @@
-package domain;
+package domain.game;
 
-import org.assertj.core.api.Assertions;
+import domain.user.Users;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ResultsTest {
 
@@ -14,7 +14,8 @@ class ResultsTest {
     @DisplayName("Result 수는 사용자 수와 일치해야 한다.")
     void equalToPersonCount() {
         //given
-        int personCount = 3;
+        List<String> usernames = List.of("honux", "crong", "jk");
+        Users users = new Users(usernames);
         List<Result> prizes = List.of(
                 new Result("꽝"),
                 new Result("5000"),
@@ -22,8 +23,7 @@ class ResultsTest {
                 new Result("3000"));
         //when
         //then
-        assertThatThrownBy(() -> Results.of(prizes, personCount))
+        assertThatThrownBy(() -> Results.of(prizes, users))
                 .isInstanceOf(IllegalArgumentException.class);
     }
-
 }
